part1: initial dual-time implementation
The dual-time seems to work and produces reasonable visualizations.
Still, it does not pass the reference test. The H values seem to be much
much lower than the reference. Also, the Xi_g values are not correct,
which is weird because with the previous test.bson the teacher gave us,
the Xi_g where the same.

Moreover, update_halo! is not working.